Move away from tflite-rs crate complately
Now that we have a functional hotg-runecoral library we are moving away
from tflite-rs crate completely.

TODO: CI for All the target architectures we support
