+++ bindings/native/examples/Makefile
@@ -5,7 +5,7 @@ TARGET_DIR := ${PROJECT_ROOT}/target
 RUNE_HEADER_FILE := ${TARGET_DIR}/rune.h
 RUST_SOURCE := $(shell find ${CRATE_DIR} -iname *.rs)
 RUST_LIB_DIR := ${TARGET_DIR}/debug
-RUST_FEATURES := wasmer-runtime,tflite
+RUST_FEATURES := wasmer-runtime
 RUNE_LIB := "${RUST_LIB_DIR}/librune_native.so"
 CXXFLAGS := -std=c++20 -g -Wl,-rpath,${RUNE_LIB}
 LIBS := -ldl -lpthread -lstdc++
