+++ images/runicos-base/runtime/Cargo.toml
@@ -19,17 +19,16 @@ hotg-rune-core = { path = "../../../crates/rune-core", version = "^0.7.1-dev", f
 hotg-rune-runtime = { path = "../../../crates/runtime", version = "^0.7.1-dev", features = ["builtins"] }
 hotg-rune-wasmer-runtime = { path = "../../../crates/wasmer-runtime", version = "^0.7.1-dev", optional = true }
 hotg-rune-wasm3-runtime = { path = "../../../crates/wasm3-runtime", version = "^0.7.1-dev", optional = true }
+hotg-runecoral = "0.3.0"
 log = "0.4.14"
 rand = "0.8.4"
 serde_json = { version = "1" }
-# Note: This doesn't cross-compile due to https://github.com/boncheolgu/tflite-rs/issues/49
-tflite = { version = "0.9.5", optional = true }
 wasmer = { version = "2.0.0", optional = true }
 wasmer-vm = { version = "2.0.0", optional = true }
 wasm3 = { version = "0.3.0", optional = true }
 
 [features]
-default = ["tflite", "wasmer-runtime"]
+default = ["wasmer-runtime"]
 wasmer-runtime = ["hotg-rune-wasmer-runtime", "wasmer", "wasmer-vm"]
 wasm3-runtime = ["hotg-rune-wasm3-runtime", "wasm3"]
